Add shell command media source drivers
New drivers to allow the use of shell command output
as a video or audio source.
Added tests and example code for this driver.
This can be used in cases where command line programs
support media sources that pion/mediadevices doesn't yet
(EG: libcamera based cameras), or to add filters to a video or audio
stream before pion/mediadevices using ffmpeg or gstreamer.